Problema al actualizar precios de catalogo y al elegir un producto para el detalle

+++ WebApp/AltivaWebApp/Services/PrecioCatalogoService.cs
@@ -44,7 +44,7 @@ namespace AltivaWebApp.Services
         {
             return reposistory.UpdatePrecio(domain);
         }
-        public bool SaveFromInventario(int idIventario)
+        public bool SaveFromInventario(long idIventario)
         {
             return reposistory.SaveFromInventario(idIventario);
         }
